refactor directories and docs rst fixed
