+++ docs/project/generate_nb_htmls.sh
@@ -3,8 +3,3 @@ for f in notebooks/*.ipynb
 do
     eval "jupyter nbconvert $f --output-dir $NOTEBOOK_HTML_DIR"
 done
-
-for f in notebooks/*/*.ipynb
-do
-    eval "jupyter nbconvert $f --output-dir $NOTEBOOK_HTML_DIR"
-done
