+++ docs/source/linear_regression/LinearRegressionClassifier.rst
