--- docs/source/models/nn.rst
@@ -1,9 +1,0 @@
-:mod:`models.nn` -- Neural Network model
-========================================
-
-.. automodule:: fromscratchtoml.models.nn
-    :synopsis: Neural Network model
-    :members:
-    :inherited-members:
-    :undoc-members:
-    :show-inheritance:
