--- docs/source/models/svm.rst
@@ -1,9 +1,0 @@
-:mod:`models.svm` -- Support Vector classification
-==================================================
-
-.. automodule:: fromscratchtoml.models.svm
-    :synopsis: Support Vector classification model
-    :members:
-    :inherited-members:
-    :undoc-members:
-    :show-inheritance:
