+++ docs/source/neural_network/nn.rst
@@ -1,0 +1,9 @@
+:mod:`neural_network.nn` -- Neural Network model
+================================================
+
+.. automodule:: fromscratchtoml.neural_network.nn
+    :synopsis: Neural Network model
+    :members:
+    :inherited-members:
+    :undoc-members:
+    :show-inheritance:
