+++ docs/source/svm/svc.rst
@@ -1,0 +1,9 @@
+:mod:`svm.svc` -- Support Vector classification
+===============================================
+
+.. automodule:: fromscratchtoml.svm.svc
+    :synopsis: Support Vector classification model
+    :members:
+    :inherited-members:
+    :undoc-members:
+    :show-inheritance:
