+++ fromscratchtoml/__init__.py
@@ -1,3 +1,3 @@
-from fromscratchtoml.models import nn # noqa:F401
+from fromscratchtoml.neural_network import nn # noqa:F401
 from fromscratchtoml.linear_regression import LinearRegressionClassifier # noqa:F401
-from fromscratchtoml.models import svm # noqa:F401
+from .base import BaseModel # noqa:F401
