--- fromscratchtoml/models/__init__.py
@@ -1,2 +1,0 @@
-from .svm import SVC # noqa:F401
-from .nn import NeuralNetwork # noqa:F401
