+++ fromscratchtoml/svm/__init__.py
@@ -1,0 +1,1 @@
+from .svc import SVC # noqa:F401
